+++ CHANGELOG.md
@@ -20,6 +20,18 @@ Enhancements:
 
 - Updated `Update-VcenterPasswordComplexity` to handle isolated VI Workload Domains. [GH-29](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/29)
 
+- Updated `Update-SsoPasswordComplexity` to handle isolated VI Workload Domains. [GH-30](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/30)
+
+- Updated `Update-SsoAccountLockout` to handle isolated VI Workload Domains. [GH-30](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/30)
+
+- Updated `Update-SsoPasswordExpiration` to handle isolated VI Workload Domains. [GH-30](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/30)
+
+- Updated `Request-SsoAccountLockout` to handle isolated VI Workload Domains. [GH-30](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/30)
+
+- Updated `Request-SsoPasswordComplexity` to handle isolated VI Workload Domains. [GH-30](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/30)
+
+- Updated `Request-SsoPasswordExpiration` to handle isolated VI Workload Domains. [GH-30](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/30)
+
 ## [v1.0.0](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/releases/tag/v1.0.0)
 
 > Release Date: 2023-04-25)